+++ velo/__init__.py
@@ -1,3 +1,3 @@
 """Mouse Input overlay for OBS. Made for aim trainers and FPS games."""
 
-__version__ = "1.1.5"
+__version__ = "1.2.0"